+++ src/main/kotlin/com/atlassian/performance/tools/infrastructure/api/jira/JiraNodeConfig.kt
@@ -6,7 +6,6 @@ import com.atlassian.performance.tools.infrastructure.api.jvm.jmx.DisabledRemote
 import com.atlassian.performance.tools.infrastructure.api.jvm.jmx.RemoteJmx
 import com.atlassian.performance.tools.infrastructure.api.splunk.DisabledSplunkForwarder
 import com.atlassian.performance.tools.infrastructure.api.splunk.SplunkForwarder
-import java.time.Duration
 
 class JiraNodeConfig(
     val name: String,
@@ -27,27 +26,6 @@ class JiraNodeConfig(
         jvmArgs = jvmArgs,
         splunkForwarder = DisabledSplunkForwarder(),
         launchTimeouts = launchTimeouts
-    )
-
-    @Deprecated(message = "Use the other constructor")
-    constructor(
-        name: String = "jira-node",
-        debug: JvmDebug = DisabledJvmDebug(),
-        remoteJmx: RemoteJmx = DisabledRemoteJmx(),
-        jvmArgs: JiraJvmArgs = JiraJvmArgs(),
-        splunkForwarder: SplunkForwarder = DisabledSplunkForwarder()
-    ) : this(
-        name = name,
-        debug = debug,
-        remoteJmx = remoteJmx,
-        jvmArgs = jvmArgs,
-        splunkForwarder = splunkForwarder,
-        launchTimeouts = JiraLaunchTimeouts(
-            offlineTimeout = Duration.ofMinutes(8),
-            initTimeout = Duration.ofMinutes(4),
-            upgradeTimeout = Duration.ofMinutes(8),
-            unresponsivenessTimeout = Duration.ofMinutes(4)
-        )
     )
 
     fun clone(